Fix the octree raycaster
We have changed the interface of the octree lib such that it reasons in
voxel space. This was not reflected in the raycaster. The fix was to
discretise the ray when interpolating the volume. We should change this
to perform the whole raycast in voxel space rather than metric space.