Add generic interpolator to octree class
We introduce a generic interpolator in the octree public interface. It
takes two arguments: the desired interpolated 3D position and a
meta-function which returns the field to be interpolated. An example
of its usage is shown in Octree::raycast where the field selector
funciton is implemented as a c++14 lambda with auto arguments. The
field selector works by taking as input a packed value of the field
type, either in its compute_type or stored_type (thanks to c++14 auto
lambda parameters) and returns the element of the packed struct on
which to perform the interpolation. We use the auto type as
interpolation can be performed either on the in memory representation
(i.e. stored_type) or in its compute representation (i.e.
compute_type).