Use uint64_t for morton utils
Uniformed the I/O types of morton utils to uint64_bit. However if in
the KFusion we use volume resolution > 2048 and volume size > 20m
things collaps pretty quickly. My guess is that something dodgy with
implicit casting is going on. We should recompile with -Wconversion and
get rid of ALL the warnings. To reproduce the bug:
 ./build/kfusion-qt-openmp --compute-size-ratio 2 --fps 0 --block-read
False --input-file
/home/ev314/projects/datasets/rgbd_dataset_freiburg2_desk/scene.raw
--icp-threshold 1e-03 --log-file ./results/2017-07-24_14:48:14/2 --mu
0.1 --init-pose 0.5,0.5,0.2 --integration-rate 1 --volume-size 30 -B 8
--tracking-rate 1 --volume-resolution 4096 --pyramid-levels 10,5,4
--rendering-rate 1 -k 580.8,581.8,308.8,253.0 -a
0.6529,-0.5483,0.3248,-0.4095