Allocate with inverse dir and adaptive step
Given a 3D input vertex and the high frequency band around it (e.g. 2
times the truncation bandwidth for a TSDF field or 3 times std dev for
OFusion), march the corresponding ray from the far end towards the
camera frame and collect all the intersected octant. The marching step
changes dinamically based on the distance travelled.