Refactor out getChildFromCode from octree class
getChildFromCode computes the the child id of the current octant and was
part of the private methods of the Octree class. We are refactoring it
as a free function child_id(octant, level, max_level) to unbind it from
a specific octree instance and in such a way that it can be used in
other contexts.

The second argument is not strictly necessary as the
level of octant is embedded in the key. However for compliance with the
existing usage in the Octree class we leave it there for the moment.
This will be consolidated when issue #33 will be addressed.