Wrap coding/encoding in keyops namespace
We provide new functions (encode, decode, code and level) to
respectively encode and decode octants morton numbers. The first two
functions respectively pack and unpack the morton id of an octant,
appending or stripping it's level. The last two instead simply retrieve
the locational part of a code and its level in the tree. Notice that
level is always relative to the tree in which the octant lives.