Split raycasting from ray bounds computation
Ray-cast is now split in two functions, one which performs the actual
ray marching and one that pre-computes the iteration boundaries along
the ray. In this way one can possibly bypass the bound compuation and
directly call the ray-marching routing with pre-computed bounds (i.e.
provided by the octree ray-iterator).